+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallEvent.swift
@@ -671,6 +671,11 @@ enum InternalSuperwallEvent {
         var params = paywallInfo.audienceFilterParams()
         if let product = product {
           params["abandoned_product_id"] = product.productIdentifier
+          if !product.period.isEmpty {
+            for (key, value) in product.attributes where key != "identifier" {
+              params["abandoned_product_\(key.camelCaseToSnakeCase())"] = value
+            }
+          }
         }
         return params
       default: